install services only for manual install and handle uninstall

+++ node/windows-packaging/CalicoWindows/libs/calico/calico.psm1
@@ -174,9 +174,16 @@ function Install-CNIPlugin()
 function Remove-CNIPlugin()
 {
     $cniConfFile = $env:CNI_CONF_DIR + "\" + $env:CNI_CONF_FILENAME
-    Write-Host "Removing $cniConfFile and Calico binaries."
-    rm $cniConfFile
-    rm "$env:CNI_BIN_DIR/calico*.exe"
+    if (Test-Path $cniConfFile) {
+        Write-Host "Removing Calico CNI conf file at $cniConfFile ..."
+        rm $cniConfFile
+    }
+
+    $cniBinPath = "$env:CNI_BIN_DIR/calico*.exe"
+    if (Test-Path $cniBinPath) {
+        Write-Host "Removing Calico CNI binaries at $cniBinPath ..."
+        rm $cniBinPath
+    }
 }
 
 function Install-NodeService()
